wk2107 | bringup | unit test & benchmarking
