+++ frame/collective/src/lib.rs
@@ -43,12 +43,11 @@
 #![recursion_limit = "128"]
 
 #[macro_use]
-// mod tests;
-// mod mock;
-// mod benchmarking;
+mod tests;
+mod mock;
+mod benchmarking;
 pub mod weights;
 
-use sp_std::if_std;
 use sp_std::{prelude::*, result};
 use sp_core::u32_trait::Value as U32;
 use sp_io::storage;
@@ -66,14 +65,37 @@ use frame_support::{
 #[cfg(feature = "std")]
 use frame_support::traits::GenesisBuild;
 
-// use sp_runtime::{RuntimeDebug, traits::Hash};
-use sp_runtime::{ RuntimeDebug, traits::{
-	Hash,
-}};
+use sp_runtime::{ RuntimeDebug, traits::{Hash}};
 use frame_system::{self as system};
+
 pub use weights::WeightInfo;
 
 pub use pallet::*;
+
+/// Simple index type for proposal counting.
+pub type ProposalIndex = u32;
+
+/// A number of members.
+///
+/// This also serves as a number of voting members, and since for motions, each member may
+/// vote exactly once, therefore also the number of votes for any given motion.
+pub type MemberCount = u32;
+
+/// Default voting strategy when a member is inactive.
+pub trait DefaultVote {
+	/// Get the default voting strategy, given:
+	///
+	/// - Whether the prime member voted Aye.
+	/// - Raw number of yes votes.
+	/// - Raw number of no votes.
+	/// - Total number of member count.
+	fn default_vote(
+		prime_vote: Option<bool>,
+		yes_votes: MemberCount,
+		no_votes: MemberCount,
+		len: MemberCount,
+	) -> bool;
+}
 
 #[frame_support::pallet]
 pub mod pallet {
@@ -97,9 +119,11 @@ pub mod pallet {
 		type Event: From<Event<Self, I>> + IsType<<Self as frame_system::Config>::Event>;
 
 		/// The time-out for council motions.
+		#[pallet::constant]
 		type MotionDuration: Get<Self::BlockNumber>;
 
 		/// Maximum number of proposals allowed to be active in parallel.
+		#[pallet::constant]
 		type MaxProposals: Get<ProposalIndex>;
 
 		/// The maximum number of members supported by the pallet. Used for weight estimation.
@@ -107,6 +131,7 @@ pub mod pallet {
 		/// NOTE:
 		/// + Benchmarks will need to be re-run and weights adjusted if this changes.
 		/// + This pallet assumes that dependents keep to the limit without enforcing it.
+		#[pallet::constant]
 		type MaxMembers: Get<MemberCount>;
 
 		/// Default vote strategy of this collective.
@@ -157,7 +182,7 @@ pub mod pallet {
 			ensure_root(origin)?;
 			if new_members.len() > T::MaxMembers::get() as usize {
 				debug::error!(
-					"New members count exceeds maximum amount of members expected. (expected: {}, actual: {})",
+					"New members count exceeds MAX members expected.(Expct: {}, Act: {})",
 					T::MaxMembers::get(),
 					new_members.len()
 				);
@@ -166,7 +191,7 @@ pub mod pallet {
 			let old = Members::<T, I>::get();
 			if old.len() > old_count as usize {
 				debug::warn!(
-					"Wrong count used to estimate set_members weight. (expected: {}, actual: {})",
+					"Wrong count used to estimate set_members weight.(Expct: {}, Act: {})",
 					old_count,
 					old.len()
 				);
@@ -209,7 +234,7 @@ pub mod pallet {
 				Event::MemberExecuted(proposal_hash, result.map(|_| ()).map_err(|e| e.error))
 			);
 
-			Ok(get_result_weight(result).map(|w| {
+			Ok(Self::get_result_weight(result).map(|w| {
 				T::WeightInfo::execute(
 					proposal_len as u32,  // B
 					members.len() as u32, // M
@@ -224,26 +249,6 @@ pub mod pallet {
 		/// `threshold` determines whether `proposal` is executed directly (`threshold < 2`)
 		/// or put up for voting.
 		///
-		/// # <weight>
-		/// ## Weight
-		/// - `O(B + M + P1)` or `O(B + M + P2)` where:
-		///   - `B` is `proposal` size in bytes (length-fee-bounded)
-		///   - `M` is members-count (code- and governance-bounded)
-		///   - branching is influenced by `threshold` where:
-		///     - `P1` is proposal execution complexity (`threshold < 2`)
-		///     - `P2` is proposals-count (code-bounded) (`threshold >= 2`)
-		/// - DB:
-		///   - 1 storage read `is_member` (codec `O(M)`)
-		///   - 1 storage read `ProposalOf::contains_key` (codec `O(1)`)
-		///   - DB accesses influenced by `threshold`:
-		///     - EITHER storage accesses done by `proposal` (`threshold < 2`)
-		///     - OR proposal insertion (`threshold <= 2`)
-		///       - 1 storage mutation `Proposals` (codec `O(P2)`)
-		///       - 1 storage mutation `ProposalCount` (codec `O(1)`)
-		///       - 1 storage write `ProposalOf` (codec `O(B)`)
-		///       - 1 storage write `Voting` (codec `O(M)`)
-		///   - 1 event
-		/// # </weight>
 		#[pallet::weight(
 			if *threshold < 2 {
 				T::WeightInfo::propose_execute(
@@ -271,7 +276,11 @@ pub mod pallet {
 			let proposal_len = proposal.using_encoded(|x| x.len());
 			ensure!(proposal_len <= length_bound as usize, <Error<T, I>>::WrongProposalLength);
 			let proposal_hash = T::Hashing::hash_of(&proposal);
-			ensure!(!<ProposalOf<T, I>>::contains_key(proposal_hash), <Error<T, I>>::DuplicateProposal);
+
+			ensure!(
+				!<ProposalOf<T, I>>::contains_key(proposal_hash),
+				<Error<T, I>>::DuplicateProposal
+			);
 
 			if threshold < 2 {
 				let seats = Self::members().len() as MemberCount;
@@ -280,7 +289,7 @@ pub mod pallet {
 					Event::Executed(proposal_hash, result.map(|_| ()).map_err(|e| e.error))
 				);
 
-				Ok(get_result_weight(result).map(|w| {
+				Ok(Self::get_result_weight(result).map(|w| {
 					T::WeightInfo::propose_execute(
 						proposal_len as u32, // B
 						members.len() as u32, // M
@@ -299,7 +308,7 @@ pub mod pallet {
 				let index = Self::proposal_count();
 
 				<ProposalCount<T, I>>::mutate(|i| *i += 1);
-				<ProposalOf<T, I>>::insert(proposal_hash, Some(*proposal));
+				<ProposalOf<T, I>>::insert(proposal_hash, *proposal);
 				let end = system::Module::<T>::block_number() + T::MotionDuration::get();
 				let votes = Votes { index, threshold, ayes: vec![who.clone()], nays: vec![], end };
 				<Voting<T, I>>::insert(proposal_hash, Some(votes));
@@ -490,7 +499,8 @@ pub mod pallet {
 			}
 		}
 
-		/// Disapprove a proposal, close, and remove it from the system, regardless of its current state.
+		/// Disapprove a proposal, close, and remove it from the system,
+		/// regardless of its current state.
 		///
 		/// Must be called by the Root origin.
 		///
@@ -539,7 +549,8 @@ pub mod pallet {
 	}
 
 	/// Old name generated by `decl_event`.
-	#[deprecated(note = "use `Event` instead")]
+	// TODO :: Have to re-check is "Note message" required ?
+	// #[deprecated(note = "use `Event` instead")]
 	pub type RawEvent<T, I = ()> = Event<T, I>;
 
 	#[pallet::error]
@@ -578,8 +589,8 @@ pub mod pallet {
 		_,
 		Identity,
 		T::Hash,
-		Option<T::Proposal>,
-		ValueQuery
+		T::Proposal,
+		OptionQuery,
 	>;
 
 	/// Votes on a given proposal, if it is ongoing.
@@ -630,7 +641,6 @@ pub mod pallet {
 			Pallet::<T, I>::initialize_members(&self.members);
 		}
 	}
-
 }
 
 #[cfg(feature = "std")]
@@ -653,61 +663,6 @@ impl<T: Config<I>, I: 'static> GenesisConfig<T, I> {
 	}
 }
 
-/// Simple index type for proposal counting.
-pub type ProposalIndex = u32;
-
-/// A number of members.
-///
-/// This also serves as a number of voting members, and since for motions, each member may
-/// vote exactly once, therefore also the number of votes for any given motion.
-pub type MemberCount = u32;
-
-/// Default voting strategy when a member is inactive.
-pub trait DefaultVote {
-	/// Get the default voting strategy, given:
-	///
-	/// - Whether the prime member voted Aye.
-	/// - Raw number of yes votes.
-	/// - Raw number of no votes.
-	/// - Total number of member count.
-	fn default_vote(
-		prime_vote: Option<bool>,
-		yes_votes: MemberCount,
-		no_votes: MemberCount,
-		len: MemberCount,
-	) -> bool;
-}
-
-/// Set the prime member's vote as the default vote.
-pub struct PrimeDefaultVote;
-
-impl DefaultVote for PrimeDefaultVote {
-	fn default_vote(
-		prime_vote: Option<bool>,
-		_yes_votes: MemberCount,
-		_no_votes: MemberCount,
-		_len: MemberCount,
-	) -> bool {
-		prime_vote.unwrap_or(false)
-	}
-}
-
-/// First see if yes vote are over majority of the whole collective. If so, set the default vote
-/// as yes. Otherwise, use the prime meber's vote as the default vote.
-pub struct MoreThanMajorityThenPrimeDefaultVote;
-
-impl DefaultVote for MoreThanMajorityThenPrimeDefaultVote {
-	fn default_vote(
-		prime_vote: Option<bool>,
-		yes_votes: MemberCount,
-		_no_votes: MemberCount,
-		len: MemberCount,
-	) -> bool {
-		let more_than_majority = yes_votes * 2 > len;
-		more_than_majority || prime_vote.unwrap_or(false)
-	}
-}
-
 /// Origin for the collective module.
 #[derive(PartialEq, Eq, Clone, RuntimeDebug, Encode, Decode)]
 pub enum RawOrigin<AccountId, I> {
@@ -718,7 +673,6 @@ pub enum RawOrigin<AccountId, I> {
 	/// Dummy to manage the fact we have instancing.
 	_Phantom(sp_std::marker::PhantomData<I>),
 }
-
 
 #[derive(PartialEq, Eq, Clone, Encode, Decode, RuntimeDebug)]
 /// Info for keeping track of a motion being voted on.
@@ -733,16 +687,6 @@ pub struct Votes<AccountId, BlockNumber> {
 	nays: Vec<AccountId>,
 	/// The hard end time of this vote.
 	end: BlockNumber,
-}
-
-/// Return the weight of a dispatch call result as an `Option`.
-///
-/// Will return the weight regardless of what the state of the result is.
-fn get_result_weight(result: DispatchResultWithPostInfo) -> Option<Weight> {
-	match result {
-		Ok(post_info) => post_info.actual_weight,
-		Err(err) => err.post_info.actual_weight,
-	}
 }
 
 impl<T: Config<I>, I: 'static>  Pallet<T, I> {
@@ -767,13 +711,7 @@ impl<T: Config<I>, I: 'static>  Pallet<T, I> {
 		let proposal_len = storage::read(&key, &mut [0; 0], 0)
 			.ok_or(Error::<T, I>::ProposalMissing)?;
 
-		// TODO :: Why len - 1  ? Why 1 byte extra from storage::read() ?.
-		let proposal_len = proposal_len - 1;
 		ensure!(proposal_len <= length_bound, Error::<T, I>::WrongProposalLength);
-
-		if_std!{
-			println!("Valid prop_len :: {:#?}", proposal_len);
-		}
 
 		let proposal = ProposalOf::<T, I>::get(hash).ok_or(Error::<T, I>::ProposalMissing)?;
 		let proposal_weight = proposal.get_dispatch_info().weight;
@@ -781,20 +719,7 @@ impl<T: Config<I>, I: 'static>  Pallet<T, I> {
 		Ok((proposal, proposal_len as usize))
 	}
 
-	/// Weight:
-	/// If `approved`:
-	/// - the weight of `proposal` preimage.
-	/// - two events deposited.
-	/// - two removals, one mutation.
-	/// - computation and i/o `O(P + L)` where:
-	///   - `P` is number of active proposals,
-	///   - `L` is the encoded length of `proposal` preimage.
-	///
-	/// If not `approved`:
-	/// - one event deposited.
-	/// Two removals, one mutation.
-	/// Computation and i/o `O(P)` where:
-	/// - `P` is number of active proposals
+	// Approve proposal
 	fn do_approve_proposal(
 		seats: MemberCount,
 		voting: Votes<T::AccountId, T::BlockNumber>,
@@ -810,12 +735,13 @@ impl<T: Config<I>, I: 'static>  Pallet<T, I> {
 			Event::Executed(proposal_hash, result.map(|_| ()).map_err(|e| e.error))
 		);
 		// default to the dispatch info weight for safety
-		let proposal_weight = get_result_weight(result).unwrap_or(dispatch_weight); // P1
+		let proposal_weight = Self::get_result_weight(result).unwrap_or(dispatch_weight); // P1
 
 		let proposal_count = Self::remove_proposal(proposal_hash);
 		(proposal_weight, proposal_count)
 	}
 
+	// disapprove proposal from the pallet
 	fn do_disapprove_proposal(proposal_hash: T::Hash) -> u32 {
 		// disapproved
 		Self::deposit_event(Event::Disapproved(proposal_hash));
@@ -833,6 +759,16 @@ impl<T: Config<I>, I: 'static>  Pallet<T, I> {
 		});
 		num_proposals as u32
 	}
+
+	/// Return the weight of a dispatch call result as an `Option`.
+	///
+	/// Will return the weight regardless of what the state of the result is.
+	fn get_result_weight(result: DispatchResultWithPostInfo) -> Option<Weight> {
+		match result {
+			Ok(post_info) => post_info.actual_weight,
+			Err(err) => err.post_info.actual_weight,
+		}
+	}
 }
 
 impl<T: Config<I>, I: 'static> ChangeMembers<T::AccountId> for Pallet<T, I> {
@@ -840,19 +776,6 @@ impl<T: Config<I>, I: 'static> ChangeMembers<T::AccountId> for Pallet<T, I> {
 	///
 	/// NOTE: Does not enforce the expected `MaxMembers` limit on the amount of members, but
 	///       the weight estimations rely on it to estimate dispatchable weight.
-	///
-	/// # <weight>
-	/// ## Weight
-	/// - `O(MP + N)`
-	///   - where `M` old-members-count (governance-bounded)
-	///   - where `N` new-members-count (governance-bounded)
-	///   - where `P` proposals-count
-	/// - DB:
-	///   - 1 storage read (codec `O(P)`) for reading the proposals
-	///   - `P` storage mutations for updating the votes (codec `O(M)`)
-	///   - 1 storage write (codec `O(N)`) for storing the new members
-	///   - 1 storage write (codec `O(1)`) for deleting the old prime
-	/// # </weight>
 	fn change_members_sorted(
 		_incoming: &[T::AccountId],
 		outgoing: &[T::AccountId],
@@ -1002,786 +925,5 @@ impl<
 	#[cfg(feature = "runtime-benchmarks")]
 	fn successful_origin() -> O {
 		O::from(RawOrigin::Members(0u32, 0u32))
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use frame_support::{Hashable, assert_ok, assert_noop, parameter_types};
-	use frame_support::{weights::{Pays}};
-	use frame_system::{self as system, EventRecord, Phase};
-	use hex_literal::hex;
-	use sp_core::H256;
-	use sp_runtime::{
-		traits::{BlakeTwo256, IdentityLookup}, testing::Header,
-		BuildStorage,
-	};
-	use crate as collective;
-
-	parameter_types! {
-		pub const BlockHashCount: u64 = 250;
-		pub const MotionDuration: u64 = 3;
-		pub const MaxProposals: u32 = 100;
-		pub const MaxMembers: u32 = 100;
-		pub BlockWeights: frame_system::limits::BlockWeights =
-			frame_system::limits::BlockWeights::simple_max(1024);
-	}
-	impl frame_system::Config for Test {
-		type BaseCallFilter = ();
-		type BlockWeights = ();
-		type BlockLength = ();
-		type DbWeight = ();
-		type Origin = Origin;
-		type Index = u64;
-		type BlockNumber = u64;
-		type Call = Call;
-		type Hash = H256;
-		type Hashing = BlakeTwo256;
-		type AccountId = u64;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type Header = Header;
-		type Event = Event;
-		type BlockHashCount = BlockHashCount;
-		type Version = ();
-		type PalletInfo = PalletInfo;
-		type AccountData = ();
-		type OnNewAccount = ();
-		type OnKilledAccount = ();
-		type SystemWeightInfo = ();
-		type SS58Prefix = ();
-	}
-	impl Config<Instance1> for Test {
-		type Origin = Origin;
-		type Proposal = Call;
-		type Event = Event;
-		type MotionDuration = MotionDuration;
-		type MaxProposals = MaxProposals;
-		type MaxMembers = MaxMembers;
-		type DefaultVote = PrimeDefaultVote;
-		type WeightInfo = ();
-	}
-	impl Config<Instance2> for Test {
-		type Origin = Origin;
-		type Proposal = Call;
-		type Event = Event;
-		type MotionDuration = MotionDuration;
-		type MaxProposals = MaxProposals;
-		type MaxMembers = MaxMembers;
-		type DefaultVote = MoreThanMajorityThenPrimeDefaultVote;
-		type WeightInfo = ();
-	}
-	impl Config for Test {
-		type Origin = Origin;
-		type Proposal = Call;
-		type Event = Event;
-		type MotionDuration = MotionDuration;
-		type MaxProposals = MaxProposals;
-		type MaxMembers = MaxMembers;
-		type DefaultVote = PrimeDefaultVote;
-		type WeightInfo = ();
-	}
-
-	pub type Block = sp_runtime::generic::Block<Header, UncheckedExtrinsic>;
-	pub type UncheckedExtrinsic = sp_runtime::generic::UncheckedExtrinsic<u32, u64, Call, ()>;
-
-	frame_support::construct_runtime!(
-		pub enum Test where
-			Block = Block,
-			NodeBlock = Block,
-			UncheckedExtrinsic = UncheckedExtrinsic
-		{
-			System: system::{Module, Call, Event<T>},
-			// Collective: collective::<Instance1>::{Module, Call, Event<T>, Config<T>},
-			Collective: collective::<Instance1>::{Module, Call, Event<T>, Origin<T>, Config<T>},
-			CollectiveMajority: collective::<Instance2>::{Module, Call, Event<T>, Origin<T>, Config<T>},
-			DefaultCollective: collective::{Module, Call, Event<T>, Origin<T>, Config<T>},
-		}
-	);
-
-	pub fn new_test_ext() -> sp_io::TestExternalities {
-		let mut ext: sp_io::TestExternalities = GenesisConfig {
-			collective_Instance1: Some(collective::GenesisConfig {
-				members: vec![1, 2, 3],
-				phantom: Default::default(),
-			}),
-			collective_Instance2: Some(collective::GenesisConfig {
-				members: vec![1, 2, 3, 4, 5],
-				phantom: Default::default(),
-			}),
-			collective: None,
-		}.build_storage().unwrap().into();
-		ext.execute_with(|| System::set_block_number(1));
-		ext
-	}
-
-	fn make_proposal(value: u64) -> Call {
-		let rval = Call::System(frame_system::Call::remark(value.encode()));
-		let proposal_len: u32 = rval.using_encoded(|p| p.len() as u32);
-		if_std!{
-			println!("prop_len :: {:#?}", proposal_len);
-		}
-		return rval
-	}
-
-	#[test]
-	fn motions_basic_environment_works() {
-		new_test_ext().execute_with(|| {
-			assert_eq!(Collective::members(), vec![1, 2, 3]);
-			assert_eq!(Collective::proposals(), Vec::<H256>::new());
-		});
-	}
-
-	#[test]
-	fn close_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let proposal_weight = proposal.get_dispatch_info().weight;
-			let hash = BlakeTwo256::hash_of(&proposal);
-
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, true));
-
-			System::set_block_number(3);
-			assert_noop!(
-				Collective::close(Origin::signed(4), hash.clone(), 0, proposal_weight, proposal_len),
-				Error::<Test, Instance1>::TooEarly
-			);
-
-			System::set_block_number(4);
-			assert_ok!(Collective::close(Origin::signed(4), hash.clone(), 0, proposal_weight, proposal_len));
-
-			let record = |event| EventRecord { phase: Phase::Initialization, event, topics: vec![] };
-			assert_eq!(System::events(), vec![
-				record(Event::collective_Instance1(RawEvent::Proposed(1, 0, hash.clone(), 3))),
-				record(Event::collective_Instance1(RawEvent::Voted(2, hash.clone(), true, 2, 0))),
-				record(Event::collective_Instance1(RawEvent::Closed(hash.clone(), 2, 1))),
-				record(Event::collective_Instance1(RawEvent::Disapproved(hash.clone())))
-			]);
-		});
-	}
-
-	#[test]
-	fn proposal_weight_limit_works_on_approve() {
-		new_test_ext().execute_with(|| {
-			let proposal = Call::Collective(crate::Call::set_members(vec![1, 2, 3], None, MaxMembers::get()));
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let proposal_weight = proposal.get_dispatch_info().weight;
-			let hash = BlakeTwo256::hash_of(&proposal);
-			// Set 1 as prime voter
-			Prime::<Test, Instance1>::set(Some(1));
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			// With 1's prime vote, this should pass
-			System::set_block_number(4);
-			assert_noop!(
-				Collective::close(Origin::signed(4), hash.clone(), 0, proposal_weight - 100, proposal_len),
-				Error::<Test, Instance1>::WrongProposalWeight
-			);
-			assert_ok!(Collective::close(Origin::signed(4), hash.clone(), 0, proposal_weight, proposal_len));
-		})
-	}
-
-	#[test]
-	fn proposal_weight_limit_ignored_on_disapprove() {
-		new_test_ext().execute_with(|| {
-			let proposal = Call::Collective(crate::Call::set_members(vec![1, 2, 3], None, MaxMembers::get()));
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let proposal_weight = proposal.get_dispatch_info().weight;
-			let hash = BlakeTwo256::hash_of(&proposal);
-
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			// No votes, this proposal wont pass
-			System::set_block_number(4);
-			assert_ok!(
-				Collective::close(Origin::signed(4), hash.clone(), 0, proposal_weight - 100, proposal_len)
-			);
-		})
-	}
-
-	#[test]
-	fn close_with_prime_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let proposal_weight = proposal.get_dispatch_info().weight;
-			let hash = BlakeTwo256::hash_of(&proposal);
-			assert_ok!(Collective::set_members(Origin::root(), vec![1, 2, 3], Some(3), MaxMembers::get()));
-
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, true));
-
-			System::set_block_number(4);
-			assert_ok!(Collective::close(Origin::signed(4), hash.clone(), 0, proposal_weight, proposal_len));
-
-			let record = |event| EventRecord { phase: Phase::Initialization, event, topics: vec![] };
-			assert_eq!(System::events(), vec![
-				record(Event::collective_Instance1(RawEvent::Proposed(1, 0, hash.clone(), 3))),
-				record(Event::collective_Instance1(RawEvent::Voted(2, hash.clone(), true, 2, 0))),
-				record(Event::collective_Instance1(RawEvent::Closed(hash.clone(), 2, 1))),
-				record(Event::collective_Instance1(RawEvent::Disapproved(hash.clone())))
-			]);
-		});
-	}
-
-	#[test]
-	fn close_with_voting_prime_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let proposal_weight = proposal.get_dispatch_info().weight;
-			let hash = BlakeTwo256::hash_of(&proposal);
-			assert_ok!(Collective::set_members(Origin::root(), vec![1, 2, 3], Some(1), MaxMembers::get()));
-
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, true));
-
-			System::set_block_number(4);
-			assert_ok!(Collective::close(Origin::signed(4), hash.clone(), 0, proposal_weight, proposal_len));
-
-			let record = |event| EventRecord { phase: Phase::Initialization, event, topics: vec![] };
-			assert_eq!(System::events(), vec![
-				record(Event::collective_Instance1(RawEvent::Proposed(1, 0, hash.clone(), 3))),
-				record(Event::collective_Instance1(RawEvent::Voted(2, hash.clone(), true, 2, 0))),
-				record(Event::collective_Instance1(RawEvent::Closed(hash.clone(), 3, 0))),
-				record(Event::collective_Instance1(RawEvent::Approved(hash.clone()))),
-				record(Event::collective_Instance1(RawEvent::Executed(hash.clone(), Err(DispatchError::BadOrigin))))
-			]);
-		});
-	}
-
-	#[test]
-	fn close_with_no_prime_but_majority_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let proposal_weight = proposal.get_dispatch_info().weight;
-			let hash = BlakeTwo256::hash_of(&proposal);
-			assert_ok!(CollectiveMajority::set_members(Origin::root(), vec![1, 2, 3, 4, 5], Some(5), MaxMembers::get()));
-
-			assert_ok!(CollectiveMajority::propose(Origin::signed(1), 5, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(CollectiveMajority::vote(Origin::signed(2), hash.clone(), 0, true));
-			assert_ok!(CollectiveMajority::vote(Origin::signed(3), hash.clone(), 0, true));
-
-			System::set_block_number(4);
-			assert_ok!(CollectiveMajority::close(Origin::signed(4), hash.clone(), 0, proposal_weight, proposal_len));
-
-			let record = |event| EventRecord { phase: Phase::Initialization, event, topics: vec![] };
-			assert_eq!(System::events(), vec![
-				record(Event::collective_Instance2(RawEvent::Proposed(1, 0, hash.clone(), 5))),
-				record(Event::collective_Instance2(RawEvent::Voted(2, hash.clone(), true, 2, 0))),
-				record(Event::collective_Instance2(RawEvent::Voted(3, hash.clone(), true, 3, 0))),
-				record(Event::collective_Instance2(RawEvent::Closed(hash.clone(), 5, 0))),
-				record(Event::collective_Instance2(RawEvent::Approved(hash.clone()))),
-				record(Event::collective_Instance2(RawEvent::Executed(hash.clone(), Err(DispatchError::BadOrigin))))
-			]);
-		});
-	}
-
-	#[test]
-	fn removal_of_old_voters_votes_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash = BlakeTwo256::hash_of(&proposal);
-			let end = 4;
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, true));
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 0, threshold: 3, ayes: vec![1, 2], nays: vec![], end })
-			);
-			Collective::change_members_sorted(&[4], &[1], &[2, 3, 4]);
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 0, threshold: 3, ayes: vec![2], nays: vec![], end })
-			);
-
-			let proposal = make_proposal(69);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash = BlakeTwo256::hash_of(&proposal);
-			assert_ok!(Collective::propose(Origin::signed(2), 2, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(Collective::vote(Origin::signed(3), hash.clone(), 1, false));
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 1, threshold: 2, ayes: vec![2], nays: vec![3], end })
-			);
-			Collective::change_members_sorted(&[], &[3], &[2, 4]);
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 1, threshold: 2, ayes: vec![2], nays: vec![], end })
-			);
-		});
-	}
-
-	#[test]
-	fn removal_of_old_voters_votes_works_with_set_members() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash = BlakeTwo256::hash_of(&proposal);
-			let end = 4;
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, true));
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 0, threshold: 3, ayes: vec![1, 2], nays: vec![], end })
-			);
-			assert_ok!(Collective::set_members(Origin::root(), vec![2, 3, 4], None, MaxMembers::get()));
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 0, threshold: 3, ayes: vec![2], nays: vec![], end })
-			);
-
-			let proposal = make_proposal(69);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash = BlakeTwo256::hash_of(&proposal);
-			assert_ok!(Collective::propose(Origin::signed(2), 2, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(Collective::vote(Origin::signed(3), hash.clone(), 1, false));
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 1, threshold: 2, ayes: vec![2], nays: vec![3], end })
-			);
-			assert_ok!(Collective::set_members(Origin::root(), vec![2, 4], None, MaxMembers::get()));
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 1, threshold: 2, ayes: vec![2], nays: vec![], end })
-			);
-		});
-	}
-
-	#[test]
-	fn propose_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash = proposal.blake2_256().into();
-			let end = 4;
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			assert_eq!(Collective::proposals(), vec![hash]);
-			assert_eq!(Collective::proposal_of(&hash), Some(proposal));
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 0, threshold: 3, ayes: vec![1], nays: vec![], end })
-			);
-
-			assert_eq!(System::events(), vec![
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Proposed(
-						1,
-						0,
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(),
-						3,
-					)),
-					topics: vec![],
-				}
-			]);
-		});
-	}
-
-	#[test]
-	fn limit_active_proposals() {
-		new_test_ext().execute_with(|| {
-			for i in 0..MaxProposals::get() {
-				let proposal = make_proposal(i as u64);
-				let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-				assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			}
-			let proposal = make_proposal(MaxProposals::get() as u64 + 1);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			assert_noop!(
-				Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len),
-				Error::<Test, Instance1>::TooManyProposals
-			);
-		})
-	}
-
-	#[test]
-	fn correct_validate_and_get_proposal() {
-		new_test_ext().execute_with(|| {
-			let proposal = Call::Collective(crate::Call::set_members(vec![1, 2, 3], None, MaxMembers::get()));
-			let length = proposal.encode().len() as u32;
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), length));
-
-			let hash = BlakeTwo256::hash_of(&proposal);
-			let weight = proposal.get_dispatch_info().weight;
-			assert_noop!(
-				Collective::validate_and_get_proposal(&BlakeTwo256::hash_of(&vec![3; 4]), length, weight),
-				Error::<Test, Instance1>::ProposalMissing
-			);
-			assert_noop!(
-				Collective::validate_and_get_proposal(&hash, length - 2, weight),
-				Error::<Test, Instance1>::WrongProposalLength
-			);
-			assert_noop!(
-				Collective::validate_and_get_proposal(&hash, length, weight - 10),
-				Error::<Test, Instance1>::WrongProposalWeight
-			);
-			let res = Collective::validate_and_get_proposal(&hash, length, weight);
-			assert_ok!(res.clone());
-			let (retrieved_proposal, len) = res.unwrap();
-			assert_eq!(length as usize, len);
-			assert_eq!(proposal, retrieved_proposal);
-		})
-	}
-
-	#[test]
-	fn motions_ignoring_non_collective_proposals_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			assert_noop!(
-				Collective::propose(Origin::signed(42), 3, Box::new(proposal.clone()), proposal_len),
-				Error::<Test, Instance1>::NotMember
-			);
-		});
-	}
-
-	#[test]
-	fn motions_ignoring_non_collective_votes_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash: H256 = proposal.blake2_256().into();
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			assert_noop!(
-				Collective::vote(Origin::signed(42), hash.clone(), 0, true),
-				Error::<Test, Instance1>::NotMember,
-			);
-		});
-	}
-
-	#[test]
-	fn motions_ignoring_bad_index_collective_vote_works() {
-		new_test_ext().execute_with(|| {
-			System::set_block_number(3);
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash: H256 = proposal.blake2_256().into();
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			assert_noop!(
-				Collective::vote(Origin::signed(2), hash.clone(), 1, true),
-				Error::<Test, Instance1>::WrongIndex,
-			);
-		});
-	}
-
-	#[test]
-	fn motions_revoting_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash: H256 = proposal.blake2_256().into();
-			let end = 4;
-			assert_ok!(Collective::propose(Origin::signed(1), 2, Box::new(proposal.clone()), proposal_len));
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 0, threshold: 2, ayes: vec![1], nays: vec![], end })
-			);
-			assert_noop!(
-				Collective::vote(Origin::signed(1), hash.clone(), 0, true),
-				Error::<Test, Instance1>::DuplicateVote,
-			);
-			assert_ok!(Collective::vote(Origin::signed(1), hash.clone(), 0, false));
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 0, threshold: 2, ayes: vec![], nays: vec![1], end })
-			);
-			assert_noop!(
-				Collective::vote(Origin::signed(1), hash.clone(), 0, false),
-				Error::<Test, Instance1>::DuplicateVote,
-			);
-
-			assert_eq!(System::events(), vec![
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Proposed(
-						1,
-						0,
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(),
-						2,
-					)),
-					topics: vec![],
-				},
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Voted(
-						1,
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(),
-						false,
-						0,
-						1,
-					)),
-					topics: vec![],
-				}
-			]);
-		});
-	}
-
-	#[test]
-	fn motions_all_first_vote_free_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash: H256 = proposal.blake2_256().into();
-			let end = 4;
-			assert_ok!(
-				Collective::propose(
-					Origin::signed(1),
-					2,
-					Box::new(proposal.clone()),
-					proposal_len,
-				)
-			);
-			assert_eq!(
-				Collective::voting(&hash),
-				Some(Votes { index: 0, threshold: 2, ayes: vec![1], nays: vec![], end })
-			);
-
-			// For the motion, acc 2's first vote, expecting Ok with Pays::No.
-			let vote_rval: DispatchResultWithPostInfo = Collective::vote(
-				Origin::signed(2),
-				hash.clone(),
-				0,
-				true,
-			);
-			assert_eq!(vote_rval.unwrap().pays_fee, Pays::No);
-
-			// Duplicate vote, expecting error with Pays::Yes.
-			let vote_rval: DispatchResultWithPostInfo = Collective::vote(
-				Origin::signed(2),
-				hash.clone(),
-				0,
-				true,
-			);
-			assert_eq!(vote_rval.unwrap_err().post_info.pays_fee, Pays::Yes);
-
-			// Modifying vote, expecting ok with Pays::Yes.
-			let vote_rval: DispatchResultWithPostInfo = Collective::vote(
-				Origin::signed(2),
-				hash.clone(),
-				0,
-				false,
-			);
-			assert_eq!(vote_rval.unwrap().pays_fee, Pays::Yes);
-
-			// For the motion, acc 3's first vote, expecting Ok with Pays::No.
-			let vote_rval: DispatchResultWithPostInfo = Collective::vote(
-				Origin::signed(3),
-				hash.clone(),
-				0,
-				true,
-			);
-			assert_eq!(vote_rval.unwrap().pays_fee, Pays::No);
-
-			// acc 3 modify the vote, expecting Ok with Pays::Yes.
-			let vote_rval: DispatchResultWithPostInfo = Collective::vote(
-				Origin::signed(3),
-				hash.clone(),
-				0,
-				false,
-			);
-			assert_eq!(vote_rval.unwrap().pays_fee, Pays::Yes);
-
-			// Test close() Extrincis | Check DispatchResultWithPostInfo with Pay Info
-
-			let proposal_weight = proposal.get_dispatch_info().weight;
-			let close_rval: DispatchResultWithPostInfo = Collective::close(
-				Origin::signed(2),
-				hash.clone(),
-				0,
-				proposal_weight,
-				proposal_len,
-			);
-			assert_eq!(close_rval.unwrap().pays_fee, Pays::No);
-
-			// trying to close the proposal, which is already closed.
-			// Expecting error "ProposalMissing" with Pays::Yes
-			let close_rval: DispatchResultWithPostInfo = Collective::close(
-				Origin::signed(2),
-				hash.clone(),
-				0,
-				proposal_weight,
-				proposal_len,
-			);
-			assert_eq!(close_rval.unwrap_err().post_info.pays_fee, Pays::Yes);
-		});
-	}
-
-	#[test]
-	fn motions_reproposing_disapproved_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let proposal_weight = proposal.get_dispatch_info().weight;
-			let hash: H256 = proposal.blake2_256().into();
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, false));
-			assert_ok!(Collective::close(Origin::signed(2), hash.clone(), 0, proposal_weight, proposal_len));
-			assert_eq!(Collective::proposals(), vec![]);
-			assert_ok!(Collective::propose(Origin::signed(1), 2, Box::new(proposal.clone()), proposal_len));
-			assert_eq!(Collective::proposals(), vec![hash]);
-		});
-	}
-
-	#[test]
-	fn motions_disapproval_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let proposal_weight = proposal.get_dispatch_info().weight;
-			let hash: H256 = proposal.blake2_256().into();
-			assert_ok!(Collective::propose(Origin::signed(1), 3, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, false));
-			assert_ok!(Collective::close(Origin::signed(2), hash.clone(), 0, proposal_weight, proposal_len));
-
-			assert_eq!(System::events(), vec![
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(
-						RawEvent::Proposed(
-							1,
-							0,
-							hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(),
-							3,
-						)),
-					topics: vec![],
-				},
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Voted(
-						2,
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(),
-						false,
-						1,
-						1,
-					)),
-					topics: vec![],
-				},
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Closed(
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(), 1, 1,
-					)),
-					topics: vec![],
-				},
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Disapproved(
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(),
-					)),
-					topics: vec![],
-				}
-			]);
-		});
-	}
-
-	#[test]
-	fn motions_approval_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let proposal_weight = proposal.get_dispatch_info().weight;
-			let hash: H256 = proposal.blake2_256().into();
-			assert_ok!(Collective::propose(Origin::signed(1), 2, Box::new(proposal.clone()), proposal_len));
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, true));
-			assert_ok!(Collective::close(Origin::signed(2), hash.clone(), 0, proposal_weight, proposal_len));
-
-			assert_eq!(System::events(), vec![
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Proposed(
-						1,
-						0,
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(),
-						2,
-					)),
-					topics: vec![],
-				},
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Voted(
-						2,
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(),
-						true,
-						2,
-						0,
-					)),
-					topics: vec![],
-				},
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Closed(
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(), 2, 0,
-					)),
-					topics: vec![],
-				},
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Approved(
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(),
-					)),
-					topics: vec![],
-				},
-				EventRecord {
-					phase: Phase::Initialization,
-					event: Event::collective_Instance1(RawEvent::Executed(
-						hex!["68eea8f20b542ec656c6ac2d10435ae3bd1729efc34d1354ab85af840aad2d35"].into(),
-						Err(DispatchError::BadOrigin),
-					)),
-					topics: vec![],
-				}
-			]);
-		});
-	}
-
-	#[test]
-	fn close_disapprove_does_not_care_about_weight_or_len() {
-		// This test confirms that if you close a proposal that would be disapproved,
-		// we do not care about the proposal length or proposal weight since it will
-		// not be read from storage or executed.
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash: H256 = proposal.blake2_256().into();
-			assert_ok!(Collective::propose(Origin::signed(1), 2, Box::new(proposal.clone()), proposal_len));
-			// First we make the proposal succeed
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, true));
-			// It will not close with bad weight/len information
-			assert_noop!(
-				Collective::close(Origin::signed(2), hash.clone(), 0, 0, 0),
-				Error::<Test, Instance1>::WrongProposalLength,
-			);
-			assert_noop!(
-				Collective::close(Origin::signed(2), hash.clone(), 0, 0, proposal_len),
-				Error::<Test, Instance1>::WrongProposalWeight,
-			);
-			// Now we make the proposal fail
-			assert_ok!(Collective::vote(Origin::signed(1), hash.clone(), 0, false));
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, false));
-			// It can close even if the weight/len information is bad
-			assert_ok!(Collective::close(Origin::signed(2), hash.clone(), 0, 0, 0));
-		})
-	}
-
-	#[test]
-	fn disapprove_proposal_works() {
-		new_test_ext().execute_with(|| {
-			let proposal = make_proposal(42);
-			let proposal_len: u32 = proposal.using_encoded(|p| p.len() as u32);
-			let hash: H256 = proposal.blake2_256().into();
-			assert_ok!(Collective::propose(Origin::signed(1), 2, Box::new(proposal.clone()), proposal_len));
-			// Proposal would normally succeed
-			assert_ok!(Collective::vote(Origin::signed(2), hash.clone(), 0, true));
-			// But Root can disapprove and remove it anyway
-			assert_ok!(Collective::disapprove_proposal(Origin::root(), hash.clone()));
-			let record = |event| EventRecord { phase: Phase::Initialization, event, topics: vec![] };
-			assert_eq!(System::events(), vec![
-				record(Event::collective_Instance1(RawEvent::Proposed(1, 0, hash.clone(), 2))),
-				record(Event::collective_Instance1(RawEvent::Voted(2, hash.clone(), true, 2, 0))),
-				record(Event::collective_Instance1(RawEvent::Disapproved(hash.clone()))),
-			]);
-		})
 	}
 }
